imgui viewport displayer + gui system changes

+++ VoxelEngine_V2/engine/utilities/stringUtil.cpp
@@ -42,7 +42,7 @@ namespace Vxl
 			return filePath.substr(lastDashLoc + 1, filePath.length() - lastDashLoc);
 		}
 		// Count Characters in string
-		uint32_t countChar(const std::string& str, char c)
+		uint64_t countChar(const std::string& str, char c)
 		{
 			return std::count(str.begin(), str.end(), c);
 		}
