+++ VoxelEngine_V2/engine/utilities/stringUtil.h
@@ -28,7 +28,7 @@ namespace Vxl
 		std::string nameFromFilepath(const std::string& filePath);
 
 		// Count Characters in string
-		uint32_t countChar(const std::string& str, char c);
+		uint64_t countChar(const std::string& str, char c);
 		
 		// Remove characters from string
 		void removeChar(std::string& str, char c);
